Make utilities a module and move its modules in the main package
