+++ data_quality/main.py
@@ -4,16 +4,13 @@ from __future__ import division
 from __future__ import print_function
 from __future__ import unicode_literals
 
-import shutil
 import os
 import sys
 import io
 import json
 import click
-import collections
 from goodtables import pipeline
-from data_quality import tasks
-
+from data_quality import tasks, utilities
 
 @click.group()
 def cli():
@@ -38,8 +35,8 @@ def run(config_filepath, deploy, encoding):
     if not os.path.isabs(config['cache_dir']):
         config['cache_dir'] = os.path.join(os.path.dirname(config_filepath),
                                            config['cache_dir'])
-                                          
-    set_up_cache_dir(config['cache_dir'])
+
+    utilities.set_up_cache_dir(config['cache_dir'])
     source_filepath = os.path.join(config['data_dir'], config['source_file'])
     default_batch_options = {
         'goodtables': {
@@ -56,14 +53,14 @@ def run(config_filepath, deploy, encoding):
             }
         }
     }
-    options = deep_update(default_batch_options, config)
+    options = utilities.deep_update_dict(default_batch_options, config)
     aggregator = tasks.Aggregator(options)
 
     if deploy:
 
         def batch_handler(instance):
             aggregator.write_run()
-            assesser = tasks.AssessPerformance(config) 
+            assesser = tasks.AssessPerformance(config)
             assesser.run()
             deployer = tasks.Deploy(config)
             deployer.run()
@@ -72,7 +69,7 @@ def run(config_filepath, deploy, encoding):
 
         def batch_handler(instance):
             aggregator.write_run()
-            assesser = tasks.AssessPerformance(config) 
+            assesser = tasks.AssessPerformance(config)
             assesser.run()
 
     post_tasks = {'post_task': batch_handler, 'pipeline_post_task': aggregator.run}
@@ -94,35 +91,6 @@ def deploy(config_filepath):
 
     deployer = tasks.Deploy(config)
     deployer.run()
-
-def set_up_cache_dir(cache_dir_path):
-    """Reset /cache_dir before a new batch."""
-
-    if os.path.lexists(cache_dir_path):
-        for root, dirs, files in os.walk(cache_dir_path):
-            for file in files:
-                os.unlink(os.path.join(root, file))
-
-            for directory in dirs:
-                shutil.rmtree(os.path.join(root, directory))
-    else: 
-        raise OSError("The folder chosen as \'cache_dir\' does not exist.")
-
-def deep_update(source_dict, new_dict):
-    """Update a nested dictionary (modified in place) with another dictionary.
-
-    Args:
-        source_dict: dict to be updated
-        new_dict: dict to update with
-
-    """
-    for key, value in new_dict.items():
-        if isinstance(value, collections.Mapping) and value:
-            returned = deep_update(source_dict.get(key, {}), value)
-            source_dict[key] = returned
-        else:
-            source_dict[key] = new_dict[key]
-    return source_dict
 
 if __name__ == '__main__':
     cli()
