+++ data_quality/tasks.py
@@ -16,8 +16,7 @@ import pytz
 import re
 import json
 import dateutil
-from .utilities import compat
-from .utilities import exceptions
+from data_quality import compat, exceptions
 
 
 @contextlib.contextmanager
@@ -41,7 +40,7 @@ class Task(object):
         self.result_file = os.path.join(self.data_dir, self.config['result_file'])
         self.run_file = os.path.join(self.data_dir, self.config['run_file'])
         self.sources_file = os.path.join(self.data_dir, self.config['source_file'])
-        self.performance_file = os.path.join(self.data_dir, 
+        self.performance_file = os.path.join(self.data_dir,
                                              self.config['performance_file'])
         self.publishers_file = os.path.join(self.data_dir,
                                              self.config['publisher_file'])
@@ -91,7 +90,7 @@ class Aggregator(Task):
                 self.fetch_data(pipeline.data.stream, source)
 
     def get_lookup(self):
-        
+
         data_key = self.config['goodtables']['arguments']['batch']['data_key']
         _keys = ['id', 'publisher_id', data_key , 'period_id']
         lookup = []
@@ -166,20 +165,20 @@ class Aggregator(Task):
             rf.write('{0}\n'.format(entry))
 
         return True
-    
+
     def fetch_data(self, data_stream, source):
         """Cache the data source in the /fetched directory"""
-        
+
         data_key = self.config['goodtables']['arguments']['batch']['data_key']
         source_name = source.get('name', source[data_key].rsplit('/', 1)[-1])
         cached_file_name = os.path.join(self.cache_dir, source_name)
         data_stream.seek(0)
-        
+
         with io.open(cached_file_name, mode='w+', encoding='utf-8') as file:
             for line in data_stream:
                 file.write(line)
 
-        
+
 class AssessPerformance(Task):
 
     """A Task runner to assess and write the performance of publishers for each
@@ -190,14 +189,14 @@ class AssessPerformance(Task):
 
     def run(self):
         """Write the performance for all publishers."""
-        
+
         def format_row(row_dict, header_list):
             ordered = list(row_dict.get(key) for key in header_list)
             string_values = [str(val) for val in ordered]
             return ','.join(string_values)
 
         publisher_ids = self.get_publishers()
-        
+
         with io.open(self.performance_file, mode='w+', encoding='utf-8') as pfile:
             fieldnames = ['publisher_id', 'period_id', 'files_count', 'score', 'valid',
                           'files_count_to_date', 'score_to_date', 'valid_to_date']
@@ -355,7 +354,7 @@ class AssessPerformance(Task):
         Args:
             period_sources: sources correspoding to a certain period
         """
-    
+
         score = 0
 
         if len(period_sources) > 0:
