--- data_quality/utilities/__init__.py
@@ -1,11 +1,0 @@
-# -*- coding: utf-8 -*-
-from __future__ import absolute_import
-from __future__ import division
-from __future__ import print_function
-from __future__ import unicode_literals
-
-
-from . import exceptions
-from . import compat
-
-__all__ = ['exceptions', 'compat']
